Support for showing the new widgets for calendar and clock
This is the bridge code needed between incidence view and
incidencedatetime object to properly show the new widgets.

In the latter we will install an event filter that will
emit the proper signals when the widget gets input.

However we'll only commit this code after we manage to get
a nice animation and "background widget" for our widgets
as using a regular panel is not optimal.

svn path=/trunk/KDE/kdepim/mobile/; revision=1176122