Signal changes in KDateEdit and KTimeEdit
1 - Rename the current timeChanged() to timeEdited() and dateChanged()
to dateEdited();
2 - Update the uses of these signals across kdepim
3 - Add the new signals timeChanged() and dateChanged() and emit them in
setTime() and setDate(), respectively.

With the listed changes, KTimeEdit and KDateEdit signals will follow what
happens in Qt Edit classes (e.g. QLineEdit): a Changed signal whenever the
content changes and an Edit signal emitted when the user edits the content.

The new signals are going to be used in the new QML widgets for time/date
selection in korganizer-mobile.


svn path=/trunk/KDE/kdepim/; revision=1179162