Don't ask 3 times if invitation should be sent.
This is for agenda's DnD, i already fixed the editor last week.

I don't like adding a bool just for this, but trunk has a clean solution, so i'm not too woried. (And I won't backport trunk's refactorings to e35).

MERGE: none (works fine in trunk)

Part of kolab/issue4570.

svn path=/branches/kdepim/enterprise/kdepim/; revision=1180025