forward port SVN commit 1183946 by winterz:
Save attachments into a temporary directory and create the symlink to the real data
in that place, rather than making the link itself a temporary file name.

eg. before the link was "file_[RANDOM].txt" and now it's "attachmentsRANDOM/file.txt"

the advantage is that the base name is still "file.txt" and that is easier to
deal with in saveAs dialogs, compared to "file_[RANDOM].txt".
kolab/issue2786

svn path=/trunk/KDE/kdepim/messageviewer/; revision=1183965