Add an OTP clone ctor.
This version clones those parts of an outer/parent OTP that
make sense, but still allows overrides for some paramters, unlike the
copy ctor one. This is necessary in order to preserve things like the
link to the top level content node of the allowAsync() flag without
having to manually set those on each nested OTP. Together with the
previous commits, this makes invitation handling in signed or encrypted
mails work.

svn path=/trunk/KDE/kdepim/messageviewer/; revision=1196469