When an ItemModifyJob finishes, keep the item's revision in mLatestRevisionByItemId hash.
If a second IMJ comes in, check if the revision is lower than the one we have in the hash. Always use the heigher one.

This fixes the case where a 2nd IMJ was started, after the 1st ended, but before the ETM was updated.

This optimization doesn't ruin conflict detection, if the item was modified externally, we will stil get the conflict dialog.

svn path=/trunk/KDE/kdepim/calendarsupport/; revision=1196535