Use the whole screen while displaying attachments
The 'Flap' widget does not properly support changing on demand the
width of it's content. Because of this, it was very hard to make it
grow after the selection of a model without modifying the 'Flap' code.

To avoid regressions, I'm postponing the fix of the 'Flap' (that is used
in too many places to be changed now - for komo3) and using the default
width to display the attachments. This solves the issue of not properly
displaying the attachments, besides not being the most beautiful way of
doing it.

Before closing this bug, we'll do some polishing on the UI.

CCBUG: 256758

svn path=/trunk/KDE/kdepim/mobile/; revision=1202469