Weekly recurring to-dos weren't working at all.
KCalCore, KOrganizer and part of IncidenceEditors use dtDue as the recurrence base. Same parts
of the editors were using dtStart, and that resulted in crazy results.

Removed code that tried to calculate the active date because we already rolled that back in e35.
We need a full solution to edit individual occurrences.