Fix saving/restoring of 'Mark As' filter
We save the return value of MessageStatus::statusStr() in the config
file to restore the message status flag later on. Unfortunately statusStr()
adds a 'U' for every state that is not 'Read'. So instead of 'G' for 'Important'
'UG' will be saved. This will fail when restoring the state. We strip away
the unnecessary 'U' for now.