New method, handleIncidenceAboutToBeModified() which contains the invitation code that should be executed before the ItemModifyJob.
If the user isn't the organizer, it asks the user if he really wants to modify the incidence.

Regarding invitations, there are questions that must be made before the ItemModifyJob and
questions that must be made after. Before this commit it wasn't possible to make this distinction.

Part of kolab/issue4096.