Framework for looking up contacts directly in nepomuk in addition to going through Akonadi and LDAP.
The results are added to the end, as a new section called "Contacts
found in your data", after any hits that come in through akonadi,
from addressbooks or from LDAP. They are relatively noisy, but still
useful in situations where you are trying to find someone you don't have
in your address books but have had dealings with in the past. In a
brighter future we might replace the whole completion thing with one
based on pimo:person, once we have meta contact merging and such in
place. For now, this is a first step.

REVIEW: 104404