A more compact 2-column layout for the address book.
This simply places the collections (address books) and
the items (names) panes one above each other instead of
side by side.  There is a new "View" menu with options
for the viewing mode - the original 3-column layout, this,
or the original "Simple" layout - and also the barcode
option if this is configured.

CCBUG:291263
REVIEW:104308
GUI:
I18N: