Speed up asyncnepomukresourceretriever by removing linear search in vector
The QHash already had the information about pending requests.

The API had the issue that in theory a different set of properties could
be requested for an existing URL, though, but in practice this is not
happening. Moved the list of properties to the constructor, to make this
clearer (and faster, too).

Discussed with Volker.