A large number of fixes in this proxy, for proper handling of extra columns.
Lost much hair in the process of debugging this, though.
The worst bit was that mapSelectionToSource, whose purpose is to remove
the extra columns, was ending up with duplicates in the QItemSelection list,
which then confused KSelectionProxyModel. Very tricky.

This fixes a number of asserts, which started with the port to KIdentityProxyModel.