Fix initial state of actions (and window title).
ETM now tells us when a folder is available, and when kmail receives this for
the current folder, it updates its copy of the collection object in
mCurrentFolder, the window title, the state of the actions, etc.

Technically we could listen to dataChanged, but that's
called so often for so many different reasons, it would just slow things down.
Especially since we don't need to be notified of future changes (new email,
mail read, etc.), because we already get that from Akonadi::Monitor).

CCBUG: 259210