This is a big refactoring that extends the behavior that we should required only parts that are really needed. Before this, the needed parts were determined based on all available filters. With this refactoring, except in some manual filtering cases, only the filters that are valid for the collection where the item is are taken into account. For example, if there was a spam filtering for POP3 (that required the whole body), when a mail arrived in an online IMAP account, the full body was requested even if the filter had the option to apply only for non-IMAP accounts. Now only the enveloper will be requested for IMAP (unless there are other filters VALID for IMAP needing more).
To make the code more clear, the DBUS API for mailfilteragent is changed, but this is no problem, as the only user of that so far is KMail, which is updated. This is internal to "kdepim".
The change also removes some method overrides to avoid confusion and code duplication. Now all methods (except when we go through dbus) work on Akonadi::Item instead of item ids (that were qlonglong and not qint64 how it is defined...). This adds a little performance penalty, but the less code worths it imo.
Big refactoring: get only those parts for filtering that are really needed.