Apply patch from Torsten Grote
"Currently, the AttachmentMissingWarning at the bottom of the compose
window is only closed when attaching a file with its own button, but not
when attaching files by other means. I made showAddAttachmentDialog()
emit a fileAttached() signal which in turn deactivates the warning.
Maybe there is a better place to emit this signal e.g. when a file has
successfully been attached."

REVIEW: 107916