Fix displaying of contacts found in the addressbook.
completion->allMatches doesn't provide all the necessary search results resulting in at least some cases in some search results not being displayed.
completions->substringCompletion on the other hand always provides the right results.
I'm not sure why the code is that complex (not sure what problem it's trying to solve), nor if anything else breaks with that change, but I don't think so.