Major speedup when switching folders (part 2/2)
Implement a "fast path" which clears everything when the storage model
emits rowsRemoved(0, count-1), rather than removing rows one by one.
(which was not very slow in general, but could get very slow when
it had to shuffle the current index around every time)

This fix is dedicated to the jazz trio Niels Lan Doky, since I mostly wrote
it between the two sets of their concert in Copenhagen.